+++ src/viz/dataview/DataView.ts
@@ -3,9 +3,10 @@ import { Layer, Source } from '@/viz';
 import { Credentials } from '@/auth';
 import { Filter, ColumnFilters, SpatialFilters } from '@/viz/filters/types';
 import { AggregationType } from '@/data/operations/aggregation/';
+import { debounce } from '@/viz/utils';
 import { DataViewImpl, GetDataOptions } from './DataViewImpl';
 import { DataViewCalculation, DataViewMode } from './mode/DataViewMode';
-import { debounce, isGeoJSONSource, DataViewEvent } from './utils';
+import { isGeoJSONSource, DataViewEvent } from './utils';
 import { SQLSource, DatasetSource, DOSource } from '../source';
 import { CartoDataViewError, dataViewErrorTypes } from './DataViewError';
 import { DataViewLocal } from './mode/DataViewLocal';
