+++ src/viz/layer/Layer.ts
@@ -10,6 +10,7 @@ import { uuidv4 } from '@/core/utils/uuid';
 import { WithEvents } from '@/core/mixins/WithEvents';
 import { DatasetSource, SQLSource, GeoJSONSource, Source } from '@/viz';
 import { LegendWidgetOptions, LegendProperties } from '@/viz/legend';
+import { debounce } from '@/viz/utils';
 import { AggregatedColumn } from '../source/Source';
 import { DOLayer } from '../deck/DOLayer';
 import { getStyles, StyleProperties, Style } from '../style';
@@ -61,6 +62,12 @@ export class Layer extends WithEvents implements StyledLayer {
   );
 
   private dataState: DATA_STATES = DATA_STATES.STARTING;
+
+  /**
+   * Debounce scope to prevent multiple calls
+   * to this._sendDataEvent('onAfterRender') and this._sendDataEvent('onViewportLoad')
+   */
+  private setOptionScope: { timeoutId?: number } = {};
 
   // #endregion
 
@@ -225,7 +232,11 @@ export class Layer extends WithEvents implements StyledLayer {
         }
       },
       onAfterRender: () => {
-        this._sendDataEvent('onAfterRender');
+        debounce(
+          () => this._sendDataEvent('onAfterRender'),
+          OPTION_DEBOUNCE_DELAY,
+          this.setOptionScope
+        )();
       }
     });
 
@@ -457,13 +468,18 @@ export class Layer extends WithEvents implements StyledLayer {
     dimensions.forEach(dimension => this._source.addField(dimension));
     columns.forEach(aggregatedColumn => this._source.addAggregatedColumn(aggregatedColumn));
 
-    return this.replaceDeckLayer();
+    if (this._source.needsInitialization) {
+      this.replaceDeckLayer();
+    }
   }
 
   // TODO not public ?
   addSourceField(field: string) {
     this._source.addField(field);
-    return this.replaceDeckLayer();
+
+    if (this._source.needsInitialization) {
+      this.replaceDeckLayer();
+    }
   }
 
   /**
@@ -582,7 +598,11 @@ export class Layer extends WithEvents implements StyledLayer {
           styleProperties.onViewportLoad(...args);
         }
 
-        this._sendDataEvent('onViewportLoad');
+        debounce(
+          () => this._sendDataEvent('onViewportLoad'),
+          OPTION_DEBOUNCE_DELAY,
+          this.setOptionScope
+        )();
       },
       onClick: this._interactivity.onClick.bind(this._interactivity),
       onHover: this._interactivity.onHover.bind(this._interactivity)
@@ -695,18 +715,20 @@ export class Layer extends WithEvents implements StyledLayer {
       (isGeoJsonLayer || referer === 'onViewportLoad')
     ) {
       this.emit(LayerEvent.DATA_READY);
-      this.emit(LayerEvent.DATA_CHANGED);
+      // this.emit(LayerEvent.DATA_CHANGED);
       this.dataState = DATA_STATES.READY;
     }
 
     if (this.dataState === DATA_STATES.UPDATING || referer === 'onViewportLoad') {
-      this.emit(LayerEvent.DATA_CHANGED);
+      // this.emit(LayerEvent.DATA_CHANGED);
       this.dataState = DATA_STATES.READY;
     }
 
     if (referer === 'onViewportLoad') {
       this.emit(LayerEvent.TILES_LOADED);
     }
+
+    this.emit(LayerEvent.DATA_CHANGED);
   }
 
   /**
@@ -809,5 +831,7 @@ interface LayerPosition {
   overLayerId?: string;
   underLayerId?: string;
 }
+
+const OPTION_DEBOUNCE_DELAY = 100;
 
 // #endregion Complement
